Changed inProcess window functionalities. Results on main screen
