+++ bower.json
@@ -3,7 +3,7 @@
   "description": "",
   "main": "index.js",
   "authors": [
-    "Maria Parshina"
+    "Maria Parshina", "Joonas Aaltonen", "Atte Laakso"
   ],
   "license": "ISC",
   "moduleType": [
