+++ public/js/inProcess.js
@@ -4,6 +4,7 @@ var index = remote.require("./index.js");
 var exec = require('child_process').exec;
 var fs = require("fs");
 const app = remote.app;
+var showPoster = index.showPoster;
 
 
 // Keskeyta returns user back to home page and erases all user's input:
@@ -19,3 +20,22 @@ $('.back').on('click', function () {
 	window.location = "http://127.0.0.1:3000/";
 });
 
+// Reads log, tries to find checksum to see if whole log exists
+setInterval(readLog, 2000);
+
+function readLog() {
+	var puttyLog = fs.readFileSync('public/logs/data.txt', 'utf8');
+
+	var checksum = "CS";
+	for (var index = 0; index < puttyLog.length; index++) {
+		var sub = puttyLog.substr(index, checksum.length);
+		// If checksum exists, show results
+		if (sub === checksum)
+		{
+			app.console.log("Checksum found, ready to show results!");
+			showPoster();
+			window.location = "http://127.0.0.1:3000/results";
+			
+		}
+	}
+}
